chore(preserve): address review — drop unused helpers, shorten docs
- delete `single_z` / `from_strings` from `sum::preserve` (only used in
  tests; the one external caller in `interface.rs` is now a one-liner)
- move the three integration tests to `tests/preserve.rs` and drop the
  now-empty `sum::preserve` module
- drop the `preserve_single_z` Python helper (only used by tests; inlined
  as `_single_z` in `test_preserve.py`)
- shorten the `preserve_strings` docstrings on the Rust field and the
  Python `new()` arg; remove the `:func:preserve_single_z` cross-reference
